+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/api/methods/groupadministration/ExportChatInviteLink.java
@@ -2,7 +2,6 @@ package org.telegram.telegrambots.api.methods.groupadministration;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.type.TypeReference;
-
 import org.telegram.telegrambots.api.methods.BotApiMethod;
 import org.telegram.telegrambots.api.objects.replykeyboard.ApiResponse;
 import org.telegram.telegrambots.exceptions.TelegramApiRequestException;
@@ -11,34 +10,46 @@ import org.telegram.telegrambots.exceptions.TelegramApiValidationException;
 import java.io.IOException;
 import java.util.Objects;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 /**
  * @author Ruben Bermudez
- * @version 1.0
- * @brief Use this method for your bot to leave a group, supergroup or channel. Returns True on success.
- * @date 20 of May of 2016
+ * @version 3.1
+ * Use this method to export an invite link to a supergroup or a channel. The bot must be an administrator in the
+ * chat for this to work and must have the appropriate admin rights. Returns exported invite link as String on success.
  */
-public class LeaveChat extends BotApiMethod<Boolean> {
-    public static final String PATH = "leaveChat";
+public class ExportChatInviteLink extends BotApiMethod<String> {
+    public static final String PATH = "exportChatInviteLink";
 
     private static final String CHATID_FIELD = "chat_id";
 
     @JsonProperty(CHATID_FIELD)
     private String chatId; ///< Unique identifier for the chat to send the message to (Or username for channels)
 
-    public LeaveChat() {
+    public ExportChatInviteLink() {
         super();
+    }
+
+    public ExportChatInviteLink(String chatId) {
+        super();
+        this.chatId = checkNotNull(chatId);
+    }
+
+    public ExportChatInviteLink(Long chatId) {
+        super();
+        this.chatId = checkNotNull(chatId).toString();
     }
 
     public String getChatId() {
         return chatId;
     }
 
-    public LeaveChat setChatId(String chatId) {
+    public ExportChatInviteLink setChatId(String chatId) {
         this.chatId = chatId;
         return this;
     }
 
-    public LeaveChat setChatId(Long chatId) {
+    public ExportChatInviteLink setChatId(Long chatId) {
         Objects.requireNonNull(chatId);
         this.chatId = chatId.toString();
         return this;
@@ -50,14 +61,14 @@ public class LeaveChat extends BotApiMethod<Boolean> {
     }
 
     @Override
-    public Boolean deserializeResponse(String answer) throws TelegramApiRequestException {
+    public String deserializeResponse(String answer) throws TelegramApiRequestException {
         try {
-            ApiResponse<Boolean> result = OBJECT_MAPPER.readValue(answer,
-                    new TypeReference<ApiResponse<Boolean>>(){});
+            ApiResponse<String> result = OBJECT_MAPPER.readValue(answer,
+                    new TypeReference<ApiResponse<String>>(){});
             if (result.getOk()) {
                 return result.getResult();
             } else {
-                throw new TelegramApiRequestException("Error leaving chat", result);
+                throw new TelegramApiRequestException("Error exporting invite link", result);
             }
         } catch (IOException e) {
             throw new TelegramApiRequestException("Unable to deserialize response", e);
@@ -67,13 +78,13 @@ public class LeaveChat extends BotApiMethod<Boolean> {
     @Override
     public void validate() throws TelegramApiValidationException {
         if (chatId == null || chatId.isEmpty()) {
-            throw new TelegramApiValidationException("ChatId can't be null", this);
+            throw new TelegramApiValidationException("ChatId can't be empty", this);
         }
     }
 
     @Override
     public String toString() {
-        return "LeaveChat{" +
+        return "ExportChatInviteLink{" +
                 "chatId='" + chatId + '\'' +
                 '}';
     }